Rename icon select dialog
