+++ client/src/components/dialogs/SelectIconDialog.vue
